Problems with the chat box
This commit was made to address problems in the chat box.
Previously, it appeared to mess up a conversation.
The chat box shouldn't refresh if it gets the same arrayList of strings
twice. It should store the last options so that even if you look at your
items you can still select a chat option.